Changed wCCT accumulation from int to long

+++ CalculateSchedules.java
@@ -296,8 +296,8 @@ public class CalculateSchedules {
             }
         }
         //calculate wCCT
-        int wCCTInt = 0;
-        int wCCTEps = 0;
+        long wCCTInt = 0;
+        long wCCTEps = 0;
         double awCCT = 0;
         for (int i = 0; i < localIds.length; i++) {
             wCCTInt += finishTimes[i].numInt * originalWeights[i];
